+++ biobb_flexserv/test/unitests/test_flexserv/test_bd_run.py
@@ -7,7 +7,7 @@ class TestBDRun():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_bd_run(self):
         bd_run(properties=self.properties, **self.paths)