+++ biobb_flexserv/test/unitests/test_flexserv/test_dmd_run.py
@@ -7,7 +7,7 @@ class TestDMDRun():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_dmd_run(self):
         dmd_run(properties=self.properties, **self.paths)