+++ biobb_flexserv/test/unitests/test_flexserv/test_nma_run.py
@@ -7,9 +7,9 @@ class TestNMARun():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
-    def test_bd_run(self):
+    def test_nma_run(self):
         nma_run(properties=self.properties, **self.paths)
         assert fx.not_empty(self.paths['output_crd_path'])
         assert fx.equal(self.paths['output_crd_path'], self.paths['ref_output_crd_path'])