+++ biobb_flexserv/test/unitests/test_pcasuite/test_pcz_bfactor.py
@@ -7,7 +7,7 @@ class TestPCZbfactor():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_pcz_bfactor(self):
         pcz_bfactor(properties=self.properties, **self.paths)