+++ biobb_flexserv/test/unitests/test_pcasuite/test_pcz_collectivity.py
@@ -7,7 +7,7 @@ class TestPCZcollectivity():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_pcz_collectivity(self):
         pcz_collectivity(properties=self.properties, **self.paths)