+++ biobb_flexserv/test/unitests/test_pcasuite/test_pcz_lindemann.py
@@ -7,7 +7,7 @@ class TestPCZlindemann():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_pcz_lindemann(self):
         pcz_lindemann(properties=self.properties, **self.paths)