+++ biobb_flexserv/test/unitests/test_pcasuite/test_pcz_similarity.py
@@ -6,8 +6,8 @@ class TestPCZsimilarity():
         fx.test_setup(self, 'pcz_similarity')
 
     def teardown_class(self):
-        #fx.test_teardown(self)
-        pass
+        fx.test_teardown(self)
+        #pass
 
     def test_pcz_similarity(self):
         pcz_similarity(properties=self.properties, **self.paths)