+++ biobb_flexserv/test/unitests/test_pcasuite/test_pcz_stiffness.py
@@ -7,7 +7,7 @@ class TestPCZstiffness():
 
     def teardown_class(self):
         fx.test_teardown(self)
-        pass
+        #pass
 
     def test_pcz_stiffness(self):
         pcz_stiffness(properties=self.properties, **self.paths)